Finish off making FatFile writable
At this point we should be able to re-write files within their current
extents safely. Allocation and freeing of extents in files *should* also
work insofar as the FAT and clusters are concerned, but the fsinfo block
is unimplemented and we're not touching ctime, mtime, et al.

Also all the writing methods in FatPath are yet to be implemented.