Remove sfn_safe
It's only called from one place, and that place would be better served
just having the regex defined next to it. Also refine the internal
handling of lfn->sfn conversions; "." and ".." must be allowed inputs
now that FatDirectory is a mapping. It's the only way FatPath.mkdir can
create the initial dir components (unless we special case DirectoryEntry
containing them?)