Re-write FatPath to use new FatDirectory
File and dir creation, removal, renaming is now re-written to use the
MutableMapping style of FatDirectory. This will initially be horribly
slow (quadratic behaviour in some enumerations); to be fixed.

This change also removes the SFN from FatPath. A name is now "just" a
name and whether that refers to a file's LFN or SFN is irrelevant. Only
LFNs will be enumerated (as usual) but SFNs are still valid for
reference and manipulation. However, there's no longer an easy way to
tie a file's LFN to its equivalent SFN.