Make read-only session-scoped disk fixtures
Speeds up testing massively and stops us eating gigabytes of temporary
space for unpacking disk images. The vast majority of the time we're
using read-only images anyway via CoW mmaps, so there's no need to
duplicate the images for every test.

Still, the new read-only fixtures ensure the file modes of the (re-used)
images are read-only, and that the open file-object is read-only too,
just to try and ensure tests don't inadvertantly break other tests by
accidentally writing to anything. Further, tests must ensure they seek
the file-object explicitly, if they require it (there's no guarantee the
file-object is at the start of the file anymore as it's passed from
test to test).