Re-write a good chunk of config handling
The ConfigArgumentParser class should be the thing handling the loading
of the default configuration, because it's bound to it by the Path-typed
configuration items. It does make it a slightly janky / manual process,
to get the right order of load config, parse args, but there's now some
explicit documentation of the procedure.