Remove template from nobodd-prep
Let's assume the user can handle copying a file... Also this frees up
the user to duplicate the file how they would like (reflink or not, and
potentially block device handling in future?). Also use logging for
console output, and raise exceptions when the boot/root partition can't
be detected but we expect them to be.