Remove ThreadingMixIn
It's entirely redundant; we don't use it at all as the selector in the
main loop handles calling the server.handle_request. Yes, this means
there's only one thread dealing with *initial* incoming requests, but
every packet after that's handled by a sub-server thread so it's not a
huge limitation.