Fix reloading inherited sockets
Sockets implicitly close on garbage collection. Unfortunately the
garbage collection of the socket (owned by the socket-server) is a bit
tricky to spot in the current code organization; need to split out the
request loop into its own routine...

Anyway, the root cause was that socket's auto-close on being collected,
and because we don't "own" the fd of the socket, we need to detach it
before this happens. Moved the fd handling logic back into the server
module; it doesn't really "belong" in a generic TFTP class.

Fixes: #8